+++ awto-log-business/src/main/java/cl/awto/hastag/repository/HashtagRepository.java
@@ -4,7 +4,9 @@ import cl.awto.hastag.entities.Hashtag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface HashtagRepository extends JpaRepository<Hashtag, Integer> {
-    Hashtag findByDescription(String description);
+    List<Hashtag> findByDescription(String description);
 }