+++ awto-log-ws/src/main/java/cl/awto/hastag/services/ApiHashServices.java
@@ -1,9 +1,11 @@
 package cl.awto.hastag.services;
 
+import cl.awto.hastag.vo.RequestHashtag;
 import cl.awto.hastag.vo.RequestLogger;
-import cl.awto.hastag.vo.ResponseId;
 import org.springframework.http.ResponseEntity;
 
 public interface ApiHashServices {
-    ResponseEntity<ResponseId> createLog(RequestLogger request);
+    ResponseEntity<Integer> createLog(RequestLogger request);
+
+    ResponseEntity<Integer> update(RequestHashtag requestHashtag);
 }