+++ awto-log-ws/src/main/java/cl/awto/hastag/vo/ResponseId.java
@@ -1,14 +1,8 @@
 package cl.awto.hastag.vo;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
 import java.io.Serializable;
 
-@Getter
-@Setter
-@NoArgsConstructor
+@Deprecated
 public class ResponseId implements Serializable {
     private Integer id;
 }